BAO-CAO-PROJECT1

+++ ndt-duanbangiay/database/migrations/2024_12_27_131528_create__n_d_t__c_t__h_o_a__d_o_n_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table -> bigInteger('ndtSanPhamID') -> references('id') -> on('NDT_SAN_PHAM');
             $table -> integer('ndtSLMua');
             $table -> float('ndtDonGiaMua');
-            $table -> float('ndtThanhTien');
+            $table -> double('ndtThanhTien');
             $table -> tinyInteger('ndtTrangThai');
             $table->timestamps();
         });
